Refactor endpoints
This is a complete refactor of the API, allowing us to easily switching
from one framework to the other, in our case `lambda_http` and `axum`.

The refactor is done in a way where the endpoints would use the same
logic and same database requests.

This patch also adds a new implementation for using Axum either as a
standalone server (to facilitate local debugging) or as a lambda
function handler.

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>